# tasks/backend/GO/gremiha3/docs/swagger.yaml
definitions:
  httpserver.ItemRequest:
    properties:
      order_id:
        example: 1
        type: integer
      product_id:
        example: 1
        type: integer
      quantity:
        example: 3
        type: integer
    type: object
  httpserver.ItemResponse:
    properties:
      id:
        type: integer
      order_id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: number
    type: object
  httpserver.OrderRequest:
    properties:
      user_id:
        example: 1
        type: integer
    type: object
  httpserver.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      state_id:
        type: integer
      total_amount:
        type: number
      user_id:
        type: integer
    type: object
  httpserver.OrderStateRequest:
    properties:
      name:
        example: в обработке
        type: string
    type: object
  httpserver.OrderStateResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  httpserver.UserRequest:
    properties:
      login:
        example: cmd@cmd.ru
        type: string
      password:
        example: "123456"
        type: string
    type: object
  httpserver.UserResponse:
    properties:
      id:
        type: integer
      login:
        type: string
      password:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
  description: An Shop service API in Go using Gin framework
  title: Shop Service API
  version: "1.0"
paths:
  /item:
    get:
      description: Return item with "id" number.
      parameters:
      - default: "1"
        description: id of the item
        example: "1"
        in: query
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.ItemResponse'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Посмотреть запись в заказе по ее id.
      tags:
      - Item
    post:
      description: Add product to order.
      parameters:
      - description: Create item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/httpserver.ItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.ItemResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавить товар в заказ.
      tags:
      - Item
  /items:
    get:
      description: Return items list by order id.
      parameters:
      - default: "10"
        description: limit records on page
        example: "10"
        in: query
        name: limit
        required: true
        type: string
      - default: "0"
        description: start of record output
        example: "0"
        in: query
        name: offset
        required: true
        type: string
      - default: "1"
        description: filter by order id
        example: "1"
        in: query
        name: orderid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpserver.ItemResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: Получить список товаров в заказе.
      tags:
      - Item
  /order:
    get:
      description: Return Order with "id" number.
      parameters:
      - default: "1"
        description: id of the order
        example: "1"
        in: query
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.OrderResponse'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Посмотреть заказ по его id или по логину пользователя.
      tags:
      - Order
    post:
      description: Создание заказа для дальнейшего его заполнения.
      parameters:
      - description: Create Order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/httpserver.OrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpserver.OrderResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавить заказ.
      tags:
      - Order
  /orders:
    get:
      description: Return Orders list.
      parameters:
      - default: "10"
        description: limit records on page
        example: "10"
        in: query
        name: limit
        required: true
        type: string
      - default: "0"
        description: start of record output
        example: "0"
        in: query
        name: offset
        required: true
        type: string
      - default: "1"
        description: filter by user id
        example: "1"
        in: query
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpserver.OrderResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: Получить список всех заказов.
      tags:
      - Order
  /orderstate:
    get:
      description: Return OrderState with "id" number.
      parameters:
      - default: "1"
        description: id of the order state
        example: "1"
        in: query
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.OrderStateResponse'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Посмотреть тип статуса по его id.
      tags:
      - OrderState
    post:
      description: Создание типа статуса заказа.
      parameters:
      - description: Create Order State type
        in: body
        name: orderState
        required: true
        schema:
          $ref: '#/definitions/httpserver.OrderStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.OrderStateResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавить тип статуса заказа.
      tags:
      - OrderState
  /orderstates:
    get:
      description: Return OrderStates list.
      parameters:
      - default: "10"
        description: limit records on page
        example: "10"
        in: query
        name: limit
        required: true
        type: string
      - default: "0"
        description: start of record output
        example: "0"
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpserver.OrderStateResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: Получить список всех статусов.
      tags:
      - OrderState
...
  /signin:
    post:
      description: Sign in as an existing user.
      parameters:
      - description: 'SignIn user. Логин указывается в формате электронной почты.
          Пароль не меньше 6 символов. Роль: super или regular'
        in: body
        name: UserRequest
        required: true
        schema:
          $ref: '#/definitions/httpserver.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Авторизоваться.
      tags:
      - Auth
  /signup:
    post:
      description: Sign up a new user in the system.
      parameters:
      - description: 'Create user. Логин указывается в формате электронной почты.
          Пароль не меньше 6 символов. Роль: super или regular'
        in: body
        name: UserRequest
        required: true
        schema:
          $ref: '#/definitions/httpserver.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpserver.UserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: error-to-create-domain-user
          schema:
            type: string
      summary: Загеристрироваться.
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"